+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Adds cached/registered filters for faster query iteration with many archetypes (#378)
 - Adds batch entity removal (#396, #402)
 - Adds batch-setting entity relation targets (#406)
+- Adds batched versions of adding, removing and exchanging components (#408)
 - Makes Ark compatible with Mooncake (#405)
 
 ### Performance